extracted method
